(#11) Abstract LogicManager into an interface
Breaking Changes:
- The original `LogicManager` class has been renamed to `SceneManager`
- `SceneManager#setup` has been renamed to `SceneManager#init`
New Additions:
- `LogicManager` interface
Other Changes:
- Removed special exception checks in `Display.java` due a lack of need.
- Removed special exception checks in `Drawable.java` due a lack of need.
