+++ src/example/java/io/github/lucasstarsz/fastj/example/bullethell/Main.java
@@ -3,17 +3,17 @@ package io.github.lucasstarsz.fastj.example.bullethell;
 import io.github.lucasstarsz.fastj.engine.FastJEngine;
 import io.github.lucasstarsz.fastj.graphics.Display;
 
-import io.github.lucasstarsz.fastj.systems.control.LogicManager;
+import io.github.lucasstarsz.fastj.systems.control.SceneManager;
 
 import java.awt.Color;
 
 import io.github.lucasstarsz.fastj.example.bullethell.scenes.GameScene;
 import io.github.lucasstarsz.fastj.example.bullethell.scenes.LoseScene;
 
-public class Main extends LogicManager {
+public class Main extends SceneManager {
 
     @Override
-    public void setup(Display display) {
+    public void init(Display display) {
         GameScene gameScene = new GameScene();
         this.addScene(gameScene);
         this.setCurrentScene(gameScene);
