+++ src/main/java/io/github/lucasstarsz/fastj/graphics/Drawable.java
@@ -163,26 +163,8 @@ public abstract class Drawable extends TaggableEntity {
      * @return Boolean value that states whether the two {@code Drawable}s intersect.
      */
     public boolean collidesWith(Drawable obj) {
-        Area otherObject;
-        Area thisObject;
-
-        try {
-            otherObject = new Area(obj.collisionPath);
-        } catch (NullPointerException e) {
-            if (!FastJEngine.getLogicManager().isSwitchingScenes()) {
-                FastJEngine.error(CollisionErrorMessage, new NullPointerException("Collision path for Drawable with id: " + obj.id + " is null"));
-            }
-            return false;
-        }
-
-        try {
-            thisObject = new Area(collisionPath);
-        } catch (NullPointerException e) {
-            if (!FastJEngine.getLogicManager().isSwitchingScenes()) {
-                FastJEngine.error(CollisionErrorMessage, new NullPointerException("Collision path for Drawable with id: " + id + " is null"));
-            }
-            return false;
-        }
+        Area thisObject = new Area(collisionPath);
+        Area otherObject = new Area(obj.collisionPath);
 
         otherObject.intersect(thisObject);
         return !otherObject.isEmpty();
