+++ src/main/java/io/github/lucasstarsz/fastj/systems/control/LogicManager.java
@@ -1,355 +1,75 @@
 package io.github.lucasstarsz.fastj.systems.control;
 
-import io.github.lucasstarsz.fastj.engine.CrashMessages;
 import io.github.lucasstarsz.fastj.engine.FastJEngine;
 import io.github.lucasstarsz.fastj.graphics.Display;
 
-import java.util.Collections;
-import java.util.LinkedHashMap;
-import java.util.List;
-import java.util.Map;
+import io.github.lucasstarsz.fastj.systems.input.InputManager;
+import io.github.lucasstarsz.fastj.systems.input.keyboard.Keyboard;
+import io.github.lucasstarsz.fastj.systems.input.mouse.Mouse;
 
-/**
- * The manager which allows for control over the scenes in a game.
- * <p>
- * A {@code LogicManager} acts as a bridge between the internals of the engine, and the developer's code.
- *
- * @author Andrew Dey
- * @version 1.0.0
- */
-public abstract class LogicManager {
+import java.awt.event.InputEvent;
 
-    private final Map<String, Scene> scenes = new LinkedHashMap<>();
-    private Scene currentScene;
-    private boolean switchingScenes;
+public interface LogicManager {
 
     /**
-     * Set up the game scenes, the display, and everything in between.
+     * Initializes the logic manager.
      * <p>
      * This method is called after the engine has been set up, and the display has been created. As it is only called
      * once, it is the best place to set some initial settings that apply to the entire game.
      *
-     * @param display The {@code Display} that the game renders to.
+     * @param display The {@code Display} that the game renders to. Useful for applying display-related settings before
+     *                the game starts.
      */
-    public abstract void setup(Display display);
+    void init(Display display);
 
     /**
-     * Updates the current scene, its behaviors, and listeners.
+     * Allows the logic manager to process all pending input events.
      *
-     * @param display The {@code Display} that the game renders to.
+     * @see Keyboard
+     * @see Mouse
+     * @see InputManager
      */
-    public void update(Display display) {
-        updateCurrentScene(display);
-    }
+    void processInputEvents();
 
     /**
-     * Renders the current scene to the {@code Display}.
-     *
-     * @param display The {@code Display} that the game renders to.
-     */
-    public void render(Display display) {
-        renderCurrentScene(display);
-    }
-
-    /**
-     * Gets the currently active scene.
-     *
-     * @return Returns the currently active scene.
-     */
-    public Scene getCurrentScene() {
-        return currentScene;
-    }
-
-    /**
-     * Sets the current scene to the scene specified.
+     * Allows the logic manager to take in an input event.
      * <p>
-     * Instead of using this method to switch scenes, it is preferred that you use the {@code switchScene(String
-     * nextScene)} method.
+     * <b>FOR IMPLEMENTORS:</b> Input events taken in should be stored for later. When {@link #processInputEvents()} is
+     * called, it will process all the events stored here.
      *
-     * @param scene The scene which the current scene will be set to.
+     * @param inputEvent The event taken in.
      */
-    public void setCurrentScene(Scene scene) {
-        setCurrentScene(scene.getSceneName());
-    }
+    void receivedInputEvent(InputEvent inputEvent);
 
     /**
-     * Sets the current scene to the scene with the name specified.
+     * Allows the logic manager to update its game state once.
      * <p>
-     * Instead of using this method to switch scenes, it is preferred that you use the {@code switchScene(String
-     * nextScene)} method.
+     * The {@code FastJEngine} attempts to call this method at most {@code FastJEngine#targetUPS} times a second. This
+     * value can be changed using {@link FastJEngine#setTargetUPS(int)}.
      *
-     * @param sceneName The name of the scene which the current scene will be set to.
+     * @param display The {@code Display} that the game renders to. Useful for checking certain attributes of the
+     *                display while updating the game state.
      */
-    public void setCurrentScene(String sceneName) {
-        sceneExistenceCheck(sceneName);
-
-        currentScene = scenes.get(sceneName);
-        switchingScenes = !currentScene.isInitialized();
-    }
+    void update(Display display);
 
     /**
-     * Gets the list of all scenes in the logic manager.
-     *
-     * @return Returns the list of scenes in the logic manager.
-     */
-    public List<Scene> getScenes() {
-        return Collections.list(Collections.enumeration(scenes.values()));
-    }
-
-    /**
-     * Gets the scene with the specified scene name, if it exists.
-     *
-     * @param sceneName The name of the scene to retrieve.
-     * @return The scene, if it exists.
-     */
-    public Scene getScene(String sceneName) {
-        sceneExistenceCheck(sceneName);
-        return scenes.get(sceneName);
-    }
-
-    /**
-     * Gets the boolean that specifies whether the logic manager is currently switching scenes.
-     *
-     * @return Returns a boolean that specifies whether the logic manager is currently switching scenes.
-     */
-    public boolean isSwitchingScenes() {
-        return switchingScenes;
-    }
-
-    /**
-     * Adds the specified scene into the logic manager.
-     *
-     * @param scene The Scene object to be added.
-     */
-    public void addScene(Scene scene) {
-        sceneNameAlreadyExistsCheck(scene.getSceneName());
-        scenes.put(scene.getSceneName(), scene);
-    }
-
-    /**
-     * Removes the specified scene from the logic manager.
-     *
-     * @param scene The Scene object to be removed.
-     */
-    public void removeScene(Scene scene) {
-        removeScene(scene.getSceneName());
-    }
-
-    /**
-     * Removes a scene from the logic manager, based on the specified scene name.
-     *
-     * @param sceneName The name of the Scene to be removed.
-     */
-    public void removeScene(String sceneName) {
-        sceneExistenceCheck(sceneName);
-        scenes.remove(sceneName);
-    }
-
-    /**
-     * Switches to the scene specified, loading that scene if necessary.
+     * Allows the logic manager to render irs game's current state to the screen.
      * <p>
-     * This is the preferred method of switching from one scene to another. However, it does not unload the last scene.
-     * That has to be done by the user.
-     *
-     * @param nextSceneName The name of the next Scene to be loaded.
-     */
-    public void switchScenes(String nextSceneName) {
-        if (!scenes.containsKey(nextSceneName)) {
-            FastJEngine.error(CrashMessages.SceneError.errorMessage,
-                    new IllegalArgumentException("A scene with the name: \"" + nextSceneName + "\" hasn't been added!"));
-        }
-
-        switchingScenes = true;
-        Display display = FastJEngine.getDisplay();
-
-        Scene nextScene = scenes.get(nextSceneName);
-        if (!nextScene.isInitialized()) {
-            nextScene.load(display);
-            nextScene.initBehaviorListeners();
-            nextScene.setInitialized(true);
-        }
-        display.setBackgroundToCameraPos(nextScene.getCamera());
-
-        setCurrentScene(nextSceneName);
-        switchingScenes = false;
-    }
-
-    /** Loads the current scene, if it's not already initialized. */
-    public void loadCurrentScene() {
-        nullSceneCheck();
-
-        if (!currentScene.isInitialized()) {
-            currentScene.load(FastJEngine.getDisplay());
-            currentScene.initBehaviorListeners();
-
-            FastJEngine.getDisplay().setBackgroundToCameraPos(currentScene.getCamera());
-        }
-
-        currentScene.setInitialized(true);
-        switchingScenes = false;
-    }
-
-    /**
-     * Safely updates the current scene.
+     * The {@code FastJEngine} attempts to call this method at most {@code FastJEngine#targetFPS} times a second. This
+     * value can be changed using {@link FastJEngine#setTargetFPS(int)}.
      *
      * @param display The {@code Display} that the game renders to.
      */
-    private void updateCurrentScene(Display display) {
-        boolean[] snapshot = createSnapshot(display);
-
-        try {
-            nullSceneCheck();
-            initSceneCheck();
-
-            currentScene.update(display);
-            currentScene.updateBehaviorListeners();
-            currentScene.inputManager.fireKeysDown();
-
-        } catch (NullPointerException e) {
-            snapshotCheck(snapshot, e);
-        }
-    }
+    void render(Display display);
 
     /**
-     * Safely renders the current scene to the Display.
-     *
-     * @param display The {@code Display} that the game renders to.
-     */
-    private void renderCurrentScene(Display display) {
-        boolean[] snapshot = createSnapshot(display);
-
-        try {
-            nullSceneCheck();
-            initSceneCheck();
-
-            display.render(
-                    currentScene.drawableManager.getGameObjects(),
-                    currentScene.drawableManager.getUIElements(),
-                    currentScene.getCamera()
-            );
-
-        } catch (NullPointerException e) {
-            snapshotCheck(snapshot, e);
-        }
-    }
-
-    /**
-     * Creates a snapshot of the {@code switchingScenes} and {@code isSwitchingFullscreenState} booleans, to make sure
-     * the game doesn't crash out due to an attempt to call methods and other fields illegally.
-     *
-     * @param display The {@code Display} to get the fullscreen state from.
-     * @return An array of booleans to check through.
-     */
-    private boolean[] createSnapshot(Display display) {
-        return new boolean[]{
-                switchingScenes,
-                display.isSwitchingScreenState()
-        };
-    }
-
-    /**
-     * Checks if the logic manager was switching scenes.
+     * Resets the logic manager entirely.
      * <p>
-     * This method takes a boolean parameter, which is a record of whether the logic manager was switching scenes at the
-     * beginning of the parent method call.
+     * This method is called when the engine exits. Due to the game engine's mutability, it is preferred that all
+     * resources of the game engine are removed gracefully.
      * <p>
-     * If the logic manager was not switching scenes, this method will error out the game engine.
-     *
-     * @param snapshot Record of whether the logic manager was switching scenes at the beginning of the parent method
-     *                 call.
-     * @param e        The NullPointerException that would be used in the error call.
+     * <b>FOR IMPLEMENTORS:</b> By the end of this method call, the logic manager should have released all its
+     * resources.
      */
-    private void snapshotCheck(boolean[] snapshot, NullPointerException e) {
-        for (boolean b : snapshot) {
-            if (b) {
-                return;
-            }
-        }
-
-        FastJEngine.error(CrashMessages.SceneError.errorMessage, e);
-    }
-
-    /**
-     * Checks if the current scene is null.
-     * <p>
-     * If the current scene is null, this throws a NullPointerException that has a customized message, based on the
-     * context of the error.
-     */
-    private void nullSceneCheck() {
-        if (currentScene == null) {
-            throw new NullPointerException((scenes.size() < 1)
-                    ?
-                    "You haven't created a Scene yet, or you haven't added it to the list of scenes for the logic manager."
-                            + System.lineSeparator()
-                            + "To add a scene, use the addScene(Scene) method in your logic manager."
-                            + System.lineSeparator()
-                            + "Then, set the current scene to the scene you just added, using the setCurrentScene(Scene) method."
-
-                    :
-                    "A current scene hasn't been set."
-                            + System.lineSeparator()
-                            + "You should set the current scene, using the setCurrentScene(Scene) method from your logic manager."
-                            + System.lineSeparator()
-                            + "Scenes added: " + scenes.keySet().toString()
-            );
-        }
-    }
-
-    /**
-     * Checks if the current scene is initialized.
-     * <p>
-     * If the current scene isn't initialized, this throws a NullPointerException with a message that says so.
-     */
-    private void initSceneCheck() {
-        if (!currentScene.isInitialized()) {
-            throw new NullPointerException(
-                    "Current scene \"" + currentScene.getSceneName() + "\" isn't initialized."
-                            + System.lineSeparator()
-                            + "You should initialize the current scene, using the initCurrentScene(Display) method from your logic manager.");
-        }
-    }
-
-    /**
-     * Checks if the specified scene's name is already in use in the logic manager.
-     * <p>
-     * If the scene name is already in use, this method will error out the game engine.
-     *
-     * @param sceneName The scene name to check for.
-     */
-    private void sceneNameAlreadyExistsCheck(String sceneName) {
-        if (scenes.containsKey(sceneName)) {
-            IllegalArgumentException e = new IllegalArgumentException(
-                    "The scene name \"" + sceneName + "\" is already in use."
-                            + System.lineSeparator()
-                            + "Scenes added: " + scenes.keySet()
-            );
-
-            FastJEngine.error(CrashMessages.SceneError.errorMessage, e);
-        }
-    }
-
-    /**
-     * Checks if the specified scene name corresponds with a scene in the logic manager.
-     * <p>
-     * If the scene name doesn't correspond with any scenes in the logic manager, this method will error out the game
-     * engine.
-     *
-     * @param sceneName Name of the scene to be checked for.
-     */
-    private void sceneExistenceCheck(String sceneName) {
-        if (!scenes.containsKey(sceneName)) {
-            FastJEngine.error(CrashMessages.SceneError.errorMessage,
-                    new IllegalArgumentException("A scene with the name: \"" + sceneName + "\" hasn't been added!"));
-        }
-    }
-
-    /** Resets the logic manager. */
-    public void reset() {
-        for (Scene s : scenes.values()) {
-            if (s.isInitialized()) {
-                s.unload(FastJEngine.getDisplay());
-            }
-        }
-        scenes.clear();
-    }
+    void reset();
 }
