+++ src/test/java/unittest/mock/MockManager.java
@@ -2,10 +2,10 @@ package unittest.mock;
 
 import io.github.lucasstarsz.fastj.graphics.Display;
 
-import io.github.lucasstarsz.fastj.systems.control.LogicManager;
+import io.github.lucasstarsz.fastj.systems.control.SceneManager;
 import io.github.lucasstarsz.fastj.systems.control.Scene;
 
-public class MockManager extends LogicManager {
+public class MockManager extends SceneManager {
 
     Scene scene;
 
@@ -14,7 +14,7 @@ public class MockManager extends LogicManager {
     }
 
     @Override
-    public void setup(Display display) {
+    public void init(Display display) {
         addScene(scene);
         setCurrentScene(scene);
         loadCurrentScene();
